+++ ShowPlantProgressPlugin.cs
@@ -123,7 +123,7 @@ public class ShowPlantProgressPlugin : BaseUnityPlugin
 		int mins = (int)(remainingSeconds / 60.0);
 		remainingSeconds -= mins * 60.0;
 		int secs = (int)(remainingSeconds);
-		if (hours > 1) return $"{hours:D2}:{mins:D2}:{secs:D2}";
+		if (hours >= 1) return $"{hours:D2}:{mins:D2}:{secs:D2}";
 		return $"{mins:D2}:{secs:D2}";
 	}