+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
+## v1.5.3 (in development)
+- Time now includes hours when there is only a single hour to display
 ## v1.5.2
 - Added option to toggle percentage color
 - Show only time when show percentage is disabled and show time is enabled